add roles, rate limiting, websockets, and admin panel

+++ src/services/student.service.js
@@ -1,13 +1,13 @@
 const Student = require("../models/student.model");
 
 // Get all students
-const getAllStudents = async (userId) => {
-    return await Student.find({ user: userId });
+const getAllStudents = async (userId, isAdmin = false) => {
+    return await Student.find(isAdmin ? {} : { user: userId });
 };
 
 // Get student by ID
-const getStudentById = async (id, userId) => {
-    return await Student.findOne({ _id: id, user: userId });
+const getStudentById = async (id, userId, isAdmin = false) => {
+    return await Student.findOne(isAdmin ? { _id: id } : { _id: id, user: userId });
 };
 
 // Create student
@@ -19,16 +19,16 @@ const createStudent = async (data, userId) => {
 };
 
 // Update student
-const updateStudent = async (id, data, userId) => {
-    return await Student.findOneAndUpdate({ _id: id, user: userId }, data, {
+const updateStudent = async (id, data, userId, isAdmin = false) => {
+    return await Student.findOneAndUpdate(isAdmin ? { _id: id } : { _id: id, user: userId }, data, {
         new: true,
         runValidators: true
     });
 };
 
 // Delete student
-const deleteStudent = async (id, userId) => {
-    return await Student.findOneAndDelete({ _id: id, user: userId });
+const deleteStudent = async (id, userId, isAdmin = false) => {
+    return await Student.findOneAndDelete(isAdmin ? { _id: id } : { _id: id, user: userId });
 };
 
 module.exports = {
